fix(dev): spawn local providers via cli
