+++ packages/alchemy/src/Local/RpcServer.ts
@@ -1,82 +1,62 @@
 import { RpcSession, type RpcCompatible, type RpcTransport } from "capnweb";
+import * as Console from "effect/Console";
 import * as Context from "effect/Context";
+import * as Deferred from "effect/Deferred";
 import * as Effect from "effect/Effect";
-import * as Fiber from "effect/Fiber";
-import * as FileSystem from "effect/FileSystem";
-import * as Schedule from "effect/Schedule";
 import type * as Scope from "effect/Scope";
-import * as Lock from "./Lock.ts";
+import type { ProviderService } from "../Provider.ts";
 import {
   serializeRpcHandlers,
-  type RpcHandlerEncoders,
-  type RpcHandlers,
-} from "./RpcHandler.ts";
-import * as RpcPaths from "./RpcPaths.ts";
+  serializeRpcStreamHandler,
+  type SerializedRpc,
+} from "./RpcSerialization.ts";
 
-export class RpcServer extends Context.Service<
-  RpcServer,
-  {
-    make: <T extends RpcCompatible<T>>(
-      main: () => T,
-    ) => Effect.Effect<
-      {
-        readonly url: string;
-      },
-      never,
-      Scope.Scope
-    >;
-  }
->()("RpcServer") {}
+export class RpcServer extends Context.Service<RpcServer, never>()(
+  "RpcServer",
+) {}
 
-export const makeRpcServer = Effect.fn(function* <T extends RpcHandlers, E, R>(
-  handlersEffect: Effect.Effect<T, E, R>,
-  schema: RpcHandlerEncoders<T>,
+export type RpcProvider = {
+  [K in keyof ProviderService]: SerializedRpc<ProviderService[K]>;
+};
+
+export interface RpcApi {
+  getProvider: (key: string) => RpcProvider;
+}
+
+export const make = Effect.fnUntraced(function* (
+  serve: <T extends RpcCompatible<T>>(handlers: {
+    session: (ws: ServerWebSocketLike) => WebSocketRpcSession<T>;
+    connect: () => void;
+    disconnect: () => void;
+  }) => Effect.Effect<{ readonly url: string }, never, Scope.Scope>,
 ) {
-  const lock = yield* Lock.Lock.use((lock) => lock.acquire);
-  const fs = yield* FileSystem.FileSystem;
-  const paths = yield* RpcPaths.RpcPaths;
-
-  const server = yield* Effect.gen(function* () {
-    const heartbeat = yield* Heartbeat;
-    const handlers = yield* handlersEffect;
-    const server = yield* RpcServer.use((server) =>
-      server.make(() =>
-        Object.assign(serializeRpcHandlers(handlers, schema), {
-          heartbeat: () => Effect.runPromise(heartbeat.touch),
-          shutdown: () => Effect.runPromise(heartbeat.shutdown),
-        }),
-      ),
-    );
-    yield* fs.writeFileString(paths.url, server.url);
-    yield* Effect.addFinalizer(() =>
-      fs.readFileString(paths.url).pipe(
-        Effect.flatMap((text) =>
-          text === server.url ? fs.remove(paths.url) : Effect.void,
-        ),
-        Effect.ignore,
-      ),
-    );
-    yield* heartbeat.await;
-  }).pipe(Effect.forkScoped);
-
-  yield* Fiber.joinAll([lock, server]);
-});
-
-const Heartbeat = Effect.gen(function* () {
-  let last = Date.now();
-  const fiber = yield* Effect.suspend(() => {
-    if (Date.now() - last > 10_000) {
-      return Effect.fail({ _tag: "Timeout" } as const);
-    }
-    return Effect.void;
-  }).pipe(Effect.repeat(Schedule.spaced("4 seconds")), Effect.forkScoped);
-  return {
-    touch: Effect.sync(() => {
-      last = Date.now();
-    }),
-    shutdown: Fiber.interrupt(fiber),
-    await: Fiber.join(fiber),
+  const context = yield* Effect.context();
+  const signalConnect = yield* Deferred.make<void>();
+  const signalDisconnect = yield* Deferred.make<void>();
+  const api: RpcApi = {
+    getProvider: (key) => {
+      const provider = context.mapUnsafe.get(key);
+      if (!provider) {
+        throw new Error(`Provider "${key}" not found`);
+      }
+      const { version, stables, tail, ...handlers } =
+        provider as ProviderService;
+      return {
+        version,
+        stables,
+        tail: tail ? serializeRpcStreamHandler(tail) : undefined,
+        ...serializeRpcHandlers(handlers),
+      };
+    },
   };
+  const { url } = yield* serve({
+    session: (ws) => makeWebSocketRpcSession(ws, api),
+    connect: () => Deferred.doneUnsafe(signalConnect, Effect.void),
+    disconnect: () => Deferred.doneUnsafe(signalDisconnect, Effect.void),
+  });
+  yield* Console.log(`<ALCHEMY_RPC_ADDRESS>${url}</ALCHEMY_RPC_ADDRESS>`);
+  yield* Deferred.await(signalConnect).pipe(Effect.timeout("10 seconds")); // TODO(john): should the timeout be shorter?
+  return yield* Deferred.await(signalDisconnect) as Effect.Effect<never>;
 });
 
 interface ServerWebSocketLike {
@@ -88,12 +68,12 @@ export type WebSocketRpcSession<T extends RpcCompatible<T>> = ReturnType<
   typeof makeWebSocketRpcSession<T>
 >;
 
-export function makeWebSocketRpcSession<T extends RpcCompatible<T>>(
+function makeWebSocketRpcSession<T extends RpcCompatible<T>>(
   ws: ServerWebSocketLike,
-  main: () => T,
+  main: T,
 ) {
   const { transport, dispatch } = makeWebSocketRpcTransport(ws);
-  const session = new RpcSession(transport, main());
+  const session = new RpcSession(transport, main);
   return { session, dispatch };
 }
